Added better CRAM support for genotype samples.
Needed to pass reference to tools (samtools and pysam) when reading CRAM
files for genotype samples. Genotyper sample table now takes an optional REF
column to point to the correct reference file for each genotype sample.
This is required if reading a CRAM aligned to a reference that htslib has not yet seen and
cacehd (~/.cache/hts-ref).